+++ .nyc_output/processinfo/214f77f8-1714-4912-ab95-8734ff350daa.json
@@ -1,0 +1,1 @@
+{"parent":"e0ef7abf-5444-4beb-8c40-889fca540c2e","pid":28205,"argv":["/usr/local/bin/node","/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/node_modules/.bin/mocha","--recursive","--exit","--timeout","1000000","--require","@babel/register","src/tests/auth.test.js","-watch"],"execArgv":[],"cwd":"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team","time":1655815852427,"ppid":28204,"coverageFilename":"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/.nyc_output/214f77f8-1714-4912-ab95-8734ff350daa.json","externalId":"","uuid":"214f77f8-1714-4912-ab95-8734ff350daa","files":["/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/index.js","/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/app.js","/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/controllers/errorController.js","/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/routes/index.js","/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/routes/api/userRoutes.js","/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/controllers/authController.js","/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/models/index.js","/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/config/config.js","/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/models/user.js","/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/controllers/userController.js","/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/helpers/generatePassword.js"]}
