--- .nyc_output/processinfo/86f6df7b-6bca-4363-88b1-173319318d67.json
@@ -1,1 +1,0 @@
-{"parent":null,"pid":26404,"argv":["/usr/local/bin/node","/usr/local/bin/npm","run","test"],"execArgv":[],"cwd":"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team","time":1655814833149,"ppid":26381,"coverageFilename":"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/.nyc_output/86f6df7b-6bca-4363-88b1-173319318d67.json","externalId":"","uuid":"86f6df7b-6bca-4363-88b1-173319318d67","files":[]}
