+++ .nyc_output/processinfo/e0ef7abf-5444-4beb-8c40-889fca540c2e.json
@@ -1,0 +1,1 @@
+{"parent":null,"pid":28181,"argv":["/usr/local/bin/node","/usr/local/bin/npm","run","test"],"execArgv":[],"cwd":"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team","time":1655815851847,"ppid":28170,"coverageFilename":"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/.nyc_output/e0ef7abf-5444-4beb-8c40-889fca540c2e.json","externalId":"","uuid":"e0ef7abf-5444-4beb-8c40-889fca540c2e","files":[]}
