+++ .nyc_output/processinfo/index.json
@@ -1,1 +1,1 @@
-{"processes":{"86f6df7b-6bca-4363-88b1-173319318d67":{"parent":null,"children":[]}},"files":{},"externalIds":{}}
+{"processes":{"214f77f8-1714-4912-ab95-8734ff350daa":{"parent":"e0ef7abf-5444-4beb-8c40-889fca540c2e","children":[]},"e0ef7abf-5444-4beb-8c40-889fca540c2e":{"parent":null,"children":["214f77f8-1714-4912-ab95-8734ff350daa"]}},"files":{"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/index.js":["214f77f8-1714-4912-ab95-8734ff350daa"],"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/app.js":["214f77f8-1714-4912-ab95-8734ff350daa"],"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/controllers/errorController.js":["214f77f8-1714-4912-ab95-8734ff350daa"],"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/routes/index.js":["214f77f8-1714-4912-ab95-8734ff350daa"],"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/routes/api/userRoutes.js":["214f77f8-1714-4912-ab95-8734ff350daa"],"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/controllers/authController.js":["214f77f8-1714-4912-ab95-8734ff350daa"],"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/models/index.js":["214f77f8-1714-4912-ab95-8734ff350daa"],"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/config/config.js":["214f77f8-1714-4912-ab95-8734ff350daa"],"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/models/user.js":["214f77f8-1714-4912-ab95-8734ff350daa"],"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/controllers/userController.js":["214f77f8-1714-4912-ab95-8734ff350daa"],"/home/arthur/Desktop/Phantom App/Back-end/phantom-be-dream-team/src/helpers/generatePassword.js":["214f77f8-1714-4912-ab95-8734ff350daa"]},"externalIds":{}}
