+++ src/controllers/authController.js
@@ -1,26 +1,7 @@
-import jwt from 'jsonwebtoken';
+import createSendToken from '../helpers/createSendToken';
 import models from '../models';
 import bcrypt from 'bcryptjs';
-const signToken = (id) =>
-  jwt.sign({id}, process.env.JWT_SECRET, {
-    expiresIn: process.env.JWT_EXPIRES_IN,
-  });
 
-const createSendToken = (user, statusCode, res) => {
-  const token = signToken(user._id);
-
-  // remove the password from the output
-  user.password = undefined;
-
-  res.status(statusCode).json({
-    status: 'success',
-    token,
-    data: {
-      user,
-    },
-    message: 'Login Successful',
-  });
-};
 
 exports.login = async (req, res, next) => {
   const {email, password} = req.body;
