+++ src/helpers/createSendToken.js
@@ -1,0 +1,22 @@
+import jwt from 'jsonwebtoken';
+const signToken = (id) =>
+  jwt.sign({id}, process.env.JWT_SECRET, {
+    expiresIn: process.env.JWT_EXPIRES_IN,
+  });
+const createSendToken = (user, statusCode, res) => {
+  const token = signToken(user._id);
+
+  // remove the password from the output
+  user.password = undefined;
+
+  res.status(statusCode).json({
+    status: 'success',
+    token,
+    data: {
+      user,
+    },
+    message: 'Login Successful',
+  });
+};
+
+export default createSendToken;
